backend/requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# POSVENDELO - Dependencias Python (Backend Server)
# Versión: 2.2.0 | PostgreSQL + asyncpg
# ═══════════════════════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────────────────────
# BASE DE DATOS - PostgreSQL (REQUERIDO)
# ─────────────────────────────────────────────────────────────────────────────────
asyncpg>=0.29.0

# ─────────────────────────────────────────────────────────────────────────────────
# API Y SERVIDOR WEB
# ─────────────────────────────────────────────────────────────────────────────────
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
slowapi>=0.1.9

# ─────────────────────────────────────────────────────────────────────────────────
# HTTP Y NETWORKING
# ─────────────────────────────────────────────────────────────────────────────────
requests>=2.31.0
httpx>=0.24.0
aiofiles>=23.1.0

# ─────────────────────────────────────────────────────────────────────────────────
# SEGURIDAD Y CRIPTOGRAFÍA
# ─────────────────────────────────────────────────────────────────────────────────
cryptography>=41.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

# ─────────────────────────────────────────────────────────────────────────────────
# FISCAL Y XML - CFDI 4.0 (defusedxml obligatorio para Parsear XML en panel Fiscal)
# ─────────────────────────────────────────────────────────────────────────────────
lxml>=4.9.0
defusedxml>=0.7.1
satcfdi>=4.9.0
pyOpenSSL>=23.0.0
python-stdnum>=2.1.0

# ─────────────────────────────────────────────────────────────────────────────────
# REPORTES Y EXPORTACIÓN
# ─────────────────────────────────────────────────────────────────────────────────
openpyxl>=3.1.0
reportlab>=4.0.0

# ─────────────────────────────────────────────────────────────────────────────────
# QR CODES Y CÓDIGOS DE BARRAS
# ─────────────────────────────────────────────────────────────────────────────────
qrcode>=7.4.0

# ─────────────────────────────────────────────────────────────────────────────────
# UTILIDADES Y FECHAS
# ─────────────────────────────────────────────────────────────────────────────────
python-dateutil>=2.8.0
pytz>=2023.3
psutil>=5.9.0

# ─────────────────────────────────────────────────────────────────────────────────
# TEMPLATES Y PDF
# ─────────────────────────────────────────────────────────────────────────────────
Jinja2>=3.0.0

# ─────────────────────────────────────────────────────────────────────────────────
# TESTING
# ─────────────────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
